Fix what Matthias pointed out
