+++ scripts/2015_nips_paper/Readme.md
@@ -1,4 +1,8 @@
 ## Reproduce results of Efficient and Robust Automated Machine Learning (Feurer et al.)
+This folder contains all necessary scripts in order to reproduce the results shown in
+Figure 3 of Efficient and Robust Automated Machine Learning (Feurer et al.). The scripts
+can be modified to include different datasets, change the runtime, etc. The scripts only
+only handles classification tasks, and balanced accuracy is used as the score metric.
 
 ### 1. Creating commands.txt
 To run the experiment, first create commands.txt by running:
@@ -15,8 +19,9 @@ Run each commands in commands.txt:
 cd run
 bash run_commands.sh
 ```
-To run the commands in parallel or in a cluster,
-modify run_commands.sh.
+Each command line in commands.txt first executes model fitting, and then creating the
+single best and ensemble trajectories. Therefore, the commands can be run in parallel
+on a cluster by modifying run_commands.sh.
 
 ### 3. Plotting the results
 To plot the results, run:
@@ -24,5 +29,6 @@ To plot the results, run:
 cd plot
 bash plot_ranks.py
 ```
+
