+++ scripts/2015_nips_paper/run/run_commands.sh
@@ -1,4 +1,6 @@
 #!/usr/bin/env bash
 
-# Run all commands in commands.txt. Modify this to run on a cluster.
+# Run all commands in commands.txt. Each command line executes first the model fitting,
+# and then creates the trajectory of a single model and the ensemble. Therefore, each
+# line can be executed separately and in parallel, for example, on a cluster environment.
 cat "../commands.txt" | while read line; do eval "$line"; done
