--- scripts/2015_nips_paper/run/run_without_metalearning.py
@@ -1,86 +1,0 @@
-import argparse
-import os
-import numpy as np
-
-from autosklearn.classification import AutoSklearnClassifier
-from autosklearn.metrics import balanced_accuracy
-import openml
-
-
-def load_task(task_id):
-    """Function used for loading data."""
-    task = openml.tasks.get_task(task_id)
-    X, y = task.get_X_and_y()
-    train_indices, test_indices = task.get_train_test_split_indices()
-    X_train = X[train_indices]
-    y_train = y[train_indices]
-    X_test = X[test_indices]
-    y_test = y[test_indices]
-    dataset = openml.datasets.get_dataset(task.dataset_id)
-    _, _, cat = dataset.get_data(return_categorical_indicator=True,
-                                 target=task.target_name)
-    del _
-    del dataset
-    cat = ['categorical' if c else 'numerical' for c in cat]
-
-    unique = np.unique(y_train)
-    mapping = {unique_value: i for i, unique_value in enumerate(unique)}
-    y_train = np.array([mapping[value] for value in y_train])
-    y_test = np.array([mapping[value] for value in y_test])
-
-    return X_train, y_train, X_test, y_test, cat
-
-
-def main(working_directory, time_limit, per_run_time_limit, task_id, seed):
-    # set this to local dataset cache
-    # openml.config.cache_directory = os.path.join(working_directory, "../cache")
-
-    seed_dir = os.path.join(working_directory, str(seed))
-    try:
-        if not os.path.exists(seed_dir):
-            os.makedirs(seed_dir)
-    except Exception:
-        print("Directory {0} aleardy created.".format(seed_dir))
-
-    tmp_dir = os.path.join(seed_dir, str(task_id))
-
-    automl_arguments = {
-        'time_left_for_this_task': time_limit,
-        'per_run_time_limit': per_run_time_limit,
-        'initial_configurations_via_metalearning': 0,
-        'ensemble_size': 0,
-        'seed': seed,
-        'ml_memory_limit': 3072,
-        'resampling_strategy': 'holdout',
-        'resampling_strategy_arguments': {'train_size': 0.67},
-        'tmp_folder': tmp_dir,
-        'delete_tmp_folder_after_terminate': False,
-        'disable_evaluator_output': False,
-    }
-
-    X_train, y_train, X_test, y_test, cat = load_task(task_id)
-
-    automl = AutoSklearnClassifier(**automl_arguments)
-
-    automl.fit(X_train, y_train,
-               dataset_name=str(task_id),
-               X_test=X_test, y_test=y_test,
-               metric=balanced_accuracy)
-
-
-if __name__ == "__main__":
-    parser = argparse.ArgumentParser()
-    parser.add_argument('--working-directory', type=str, required=True)
-    parser.add_argument('--time-limit', type=int, required=True)
-    parser.add_argument('--per-run-time-limit', type=int, required=True)
-    parser.add_argument('--task-id', type=int, required=True)
-    parser.add_argument('-s', '--seed', type=int, required=True)
-
-    args = parser.parse_args()
-    working_directory = args.working_directory
-    time_limit = args.time_limit
-    per_run_time_limit = args.per_run_time_limit
-    task_id = args.task_id
-    seed = args.seed
-
-    main(working_directory, time_limit, per_run_time_limit, task_id, seed)
