--- scripts/2015_nips_paper/setup/datasets.csv
@@ -1,140 +1,0 @@
-datasets/openml/1000_bac
-datasets/openml/1002_bac
-datasets/openml/1018_bac
-datasets/openml/1019_bac
-datasets/openml/1020_bac
-datasets/openml/1021_bac
-datasets/openml/1036_bac
-datasets/openml/1040_bac
-datasets/openml/1041_bac
-datasets/openml/1049_bac
-datasets/openml/1050_bac
-datasets/openml/1053_bac
-datasets/openml/1056_bac
-datasets/openml/1067_bac
-datasets/openml/1068_bac
-datasets/openml/1069_bac
-datasets/openml/1111_bac
-datasets/openml/1112_bac
-datasets/openml/1114_bac
-datasets/openml/1116_bac
-datasets/openml/1119_bac
-datasets/openml/1120_bac
-datasets/openml/1128_bac
-datasets/openml/1130_bac
-datasets/openml/1134_bac
-datasets/openml/1138_bac
-datasets/openml/1139_bac
-datasets/openml/1142_bac
-datasets/openml/1146_bac
-datasets/openml/1161_bac
-datasets/openml/1166_bac
-datasets/openml/12_bac
-datasets/openml/14_bac
-datasets/openml/16_bac
-datasets/openml/179_bac
-datasets/openml/180_bac
-datasets/openml/181_bac
-datasets/openml/182_bac
-datasets/openml/184_bac
-datasets/openml/185_bac
-datasets/openml/18_bac
-datasets/openml/21_bac
-datasets/openml/22_bac
-datasets/openml/23_bac
-datasets/openml/24_bac
-datasets/openml/26_bac
-datasets/openml/273_bac
-datasets/openml/28_bac
-datasets/openml/293_bac
-datasets/openml/300_bac
-datasets/openml/30_bac
-datasets/openml/31_bac
-datasets/openml/32_bac
-datasets/openml/351_bac
-datasets/openml/354_bac
-datasets/openml/357_bac
-datasets/openml/36_bac
-datasets/openml/389_bac
-datasets/openml/38_bac
-datasets/openml/390_bac
-datasets/openml/391_bac
-datasets/openml/392_bac
-datasets/openml/393_bac
-datasets/openml/395_bac
-datasets/openml/396_bac
-datasets/openml/398_bac
-datasets/openml/399_bac
-datasets/openml/3_bac
-datasets/openml/401_bac
-datasets/openml/44_bac
-datasets/openml/46_bac
-datasets/openml/554_bac
-datasets/openml/57_bac
-datasets/openml/60_bac
-datasets/openml/679_bac
-datasets/openml/6_bac
-datasets/openml/715_bac
-datasets/openml/718_bac
-datasets/openml/720_bac
-datasets/openml/722_bac
-datasets/openml/723_bac
-datasets/openml/727_bac
-datasets/openml/728_bac
-datasets/openml/734_bac
-datasets/openml/735_bac
-datasets/openml/737_bac
-datasets/openml/740_bac
-datasets/openml/741_bac
-datasets/openml/743_bac
-datasets/openml/751_bac
-datasets/openml/752_bac
-datasets/openml/761_bac
-datasets/openml/772_bac
-datasets/openml/797_bac
-datasets/openml/799_bac
-datasets/openml/803_bac
-datasets/openml/806_bac
-datasets/openml/807_bac
-datasets/openml/813_bac
-datasets/openml/816_bac
-datasets/openml/819_bac
-datasets/openml/821_bac
-datasets/openml/822_bac
-datasets/openml/823_bac
-datasets/openml/833_bac
-datasets/openml/837_bac
-datasets/openml/843_bac
-datasets/openml/845_bac
-datasets/openml/846_bac
-datasets/openml/847_bac
-datasets/openml/849_bac
-datasets/openml/866_bac
-datasets/openml/871_bac
-datasets/openml/881_bac
-datasets/openml/897_bac
-datasets/openml/901_bac
-datasets/openml/903_bac
-datasets/openml/904_bac
-datasets/openml/910_bac
-datasets/openml/912_bac
-datasets/openml/913_bac
-datasets/openml/914_bac
-datasets/openml/917_bac
-datasets/openml/923_bac
-datasets/openml/930_bac
-datasets/openml/934_bac
-datasets/openml/953_bac
-datasets/openml/958_bac
-datasets/openml/959_bac
-datasets/openml/962_bac
-datasets/openml/966_bac
-datasets/openml/971_bac
-datasets/openml/976_bac
-datasets/openml/977_bac
-datasets/openml/978_bac
-datasets/openml/979_bac
-datasets/openml/980_bac
-datasets/openml/991_bac
-datasets/openml/993_bac
-datasets/openml/995_bac
